+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 **Enhancements:**
 
 - Added error message to indicate that trailing backslashes should not be used in a directory path specification. (#1166) (@hitenkoku, thanks for the suggestion from @joswr1ght)
+- Optimized the number of records to load at a time. (#1175) (@yamatosecurity)
 
 **Bug Fixes:**